+++ src/main/java/com/projet/fibonacci/modele/CalculateurFibonacci.java
@@ -1,9 +1,11 @@
 package com.projet.fibonacci.modele;
 
+import java.math.BigInteger;
+
 public interface CalculateurFibonacci {
-    long calculer(int terme);
-    default long[] calculerSuite(int nombreDeTermes) {
-        long[] suite = new long[nombreDeTermes];
+    BigInteger calculer(int terme);
+    default BigInteger[] calculerSuite(int nombreDeTermes) {
+        BigInteger[] suite = new BigInteger[nombreDeTermes];
         for (int i = 0; i < nombreDeTermes; i++) {
             suite[i] = calculer(i);
         }
